191007 | django 기본 UI 완성 (#6)
* 190930 | django web framework, first commit

* 190930 | create new project 'django-project'

* 191001 | add new app 'snapshot2'

feature: upload files to media/ with POST-FORM

* 191002 | remodelling page and fix etc

1. remodelling page: add 'admin page' link, check video/image
2. update gitignore: ignore db files
3. add scripts: runserver.sh, setupdb.sh

setupdb.sh : to setup database, just run once
runserver.sh : to run server, it will run as daemon

* 191002 | bug fix

shell script
source -> .

* 191004 | add snapshot feature

convert video to snapshots

* 191006 | implement to convert snapshots with 'caffe model'

add .gitignore: caffe model file (pose_iter_440000.caffemodel)
add coco.py: implementation of caffemodel

* 191007 | add 'snapshot3' app

version3
1. using primary key, now we can handling duplicate videos
2. implements 'clear db' feature

* 191007 | add navigator & async snapshot

1. add navigator on detail page
2. display the snapshot clicked on navigator
3. now, the function making snapshot and getting skeleton is called asynchronously

+++ src/django-project/snapshot3/views.py
@@ -4,6 +4,7 @@ from .forms import FileForm
 from .snapshot_image import snapshot
 from django.conf import settings
 import os, shutil
+from .tasks import snapshot_celery
 
 # Create your views here.
 SEC = 6  # 몇초마다 snapshot 할지 결정
@@ -20,7 +21,8 @@ def video_new(request):
             form.save()  # 파일 저장
             files = FileModel.objects.all()
             url = files[len(files)-1].file.url  # 아까 저장된 파일 url
-            save_dir = snapshot(url, SEC)  # 두번째 인자가 SEC, '몇초마다 snapshot 할지 결정'
+            # save_dir = snapshot(url, SEC)  # 두번째 인자가 SEC, '몇초마다 snapshot 할지 결정'
+            save_dir = snapshot_celery.delay(url, SEC)
             return redirect('home')
     else:
         form = FileForm()
@@ -33,7 +35,30 @@ def video_detail(request, pk):
     snapshots = os.listdir(snaps_dir)  # 전체 snapshot 리스트
     snapshots = sorted(snapshots, key=lambda x: int(x.split('.')[0]))
     snapshots = list(map(lambda x: os.path.join('snapshots', videoname, x), snapshots))  # 숫자 기준으로 정렬
+
+    # [[index, snapshot], [index, snapshot], ... ] 형태의 리스트로 변환
+    snapshots = [[idx, snapshot] for idx, snapshot in enumerate(snapshots)]
     return render(request, 'video_detail.html', {'video':video, 'snapshots':snapshots})
+
+# 클릭된 스냅샷 렌더링
+def video_snapshot(request, pk, idx):
+    idx = int(idx)
+    # video_detail 과 동일
+    video = get_object_or_404(FileModel, pk=pk)
+    videoname = os.path.basename(video.file.url)
+    snaps_dir = os.path.join(settings.SNAPS_DIR, videoname)
+    snapshots = os.listdir(snaps_dir)
+    snapshots = sorted(snapshots, key=lambda x: int(x.split('.')[0]))
+    snapshots = list(map(lambda x: os.path.join('snapshots', videoname, x), snapshots))
+    
+    dict_snapshots = {index:snapshot for index, snapshot in enumerate(snapshots)}
+
+    # [[index, snapshot], [index, snapshot], ... ] 형태의 리스트로 변환
+    snapshots = [[idx, snapshot] for idx, snapshot in enumerate(snapshots)]
+
+    snapshot = dict_snapshots[idx]  # 선택된 스냅샷
+    sec = idx*SEC
+    return render(request, 'video_detail.html', {'video':video, 'snapshots':snapshots, 'snap':snapshot, 'sec':sec})
 
 # db 내 FileModel objects
 # media/ 내 video 파일
@@ -42,8 +67,11 @@ def video_detail(request, pk):
 def clear(request):
     me = settings.MEDIA_ROOT
     sn = settings.SNAPS_DIR
-    shutil.rmtree(me)
-    shutil.rmtree(sn)
+    
+    if os.path.exists(me):
+        shutil.rmtree(me)
+    if os.path.exists(sn):
+        shutil.rmtree(sn)
     os.makedirs(me)
     os.makedirs(sn)
     FileModel.objects.all().delete()
